+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 0.1.0-beta005
 * Rework interfaces that returned `Alt`s
+* Re-enable logging
 
 ### 0.1.0-beta004
 * Better disposal of consumed responses